# nooblyjs-application-wiki/backend/swagger/openapi.full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NooblyJS Wiki Platform API",
    "description": "Unified REST API for the NooblyJS Wiki Platform (port 9101). Covers the Datasources module (`/api/*`), Wiki module (`/applications/wiki/api/*`) and Blueprint module (`/applications/blueprint/api/*`).",
    "version": "2.0.0",
    "contact": { "name": "NooblyJS Team" },
    "license": { "name": "ISC" }
  },
  "servers": [
    { "url": "/", "description": "Unified backend (port 9101)" }
  ],
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie issued after login"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Bearer token (24 hr). Obtain from POST /api/auth/token or the API tokens section of your profile."
      }
    },
...
  },
  "security": [
    { "SessionAuth": [] },
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/spaces/{id}/users/{userId}": {
      "post": {
        "tags": ["Datasources - Spaces"],
        "summary": "Add a user to a space",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "User added" } }
      },
      "delete": {
        "tags": ["Datasources - Spaces"],
        "summary": "Remove a user from a space",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "User removed" } }
      }
    },
...
  }
}